Generation_Controls/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
httpx